collections: unify ipv4 and ipv6 proto dependency
Redmine-patch-id: 5812
The following protocols are always installed in the LuCI.
* static (IPv4 and IPv6)
* none
* dhcp (only IPv4)

If IPv6 is available, the following protocols are also installed, which
have so far been combined in the luci-proto-ipv6 package.
* 464xlat
* 6in4
* 6rd
* 6to4
* dhcpv6
* dslite
* map

Not all protocols are required for ipv6. So that the protocol menu is
not overloaded, this commit changes the dependency to luci-proto-dhcpv6.
This way not all IPv6 protocols are installed, and the protocol stack
for IPv6 is the same as for IPv4. The protocols can still be installed
via the packet management opkg.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>